# projects/frontend/openapi.json
{
  "components": {
    "schemas": {
      "AuthErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "MeOk": {
        "properties": {
          "role": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "role"
        ],
        "type": "object"
      },
...
      "SessionOk": {
        "properties": {
          "role": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "role"
        ],
        "type": "object"
      },
...
      "SigninRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
      "SignupRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "orca local dev tool — docs, services, schema, MCP proxy",
    "license": {
      "name": ""
    },
    "title": "orca API",
    "version": "0.0.3-rc.34"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/me": {
      "get": {
        "operationId": "authMe",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOk"
                }
              }
            },
            "description": "Current identity"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Not signed in"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/signin": {
      "post": {
        "operationId": "authSignin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOk"
                }
              }
            },
            "description": "Signed in; session cookie set"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Invalid credentials"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/signup": {
      "post": {
        "operationId": "authSignup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOk"
                }
              }
            },
            "description": "Account created; session cookie set"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Bad request (validation)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Public sign-up disabled"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Username already taken"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
...
}